requirements.txt
aiofiles==0.3.2
appdirs==1.4.3
asyncio==3.4.3
attrs==18.1.0
autopep8==1.3.5
black==18.6b4
click==6.7
flake8==3.5.0
GraphQL-core-next==1.0.0rc2
httptools==0.0.11
mccabe==0.6.1
pep8==1.7.1
pkg-resources==0.0.0
promise==2.1
pycodestyle==2.3.1
pyflakes==1.6.0
sanic==0.7.0
six==1.11.0
toml==0.9.4
typing==3.6.4
ujson==1.35
uvloop==0.11.1
websockets==6.0
